Add errcheck to the linter.
This commit adds errcheck to the linter, and fixes _most_
instances where errors were not handled. In some cases, minor
refactoring work was done to facilitate this. In some other cases,
linting was disabled with //nolint or with a package-level ignore.

Follow-up tickets were filed for instances where linting was
disabled.

In most cases, handling errors was simple, but in some cases,
it would change the semantics of the code. I tried to preserve the
sementics of the code as faithfully as possible, while still
handling errors.

In some cases, this meant explicitly ignoring errors with _ = err.

Refs #619